+++ dataland-frontend/src/components/forms/CreateCompany.vue
@@ -233,7 +233,7 @@ import { getAllCountryCodes } from "@/utils/CountryCodeConverter";
 import { assertDefined } from "@/utils/TypeScriptUtils";
 import SuccessMessage from "@/components/messages/SuccessMessage.vue";
 import FailMessage from "@/components/messages/FailMessage.vue";
-import { checkCustomInputs } from "@/utils/validationsUtils";
+import { checkCustomInputs } from "@/utils/ValidationsUtils";
 import Tooltip from "primevue/tooltip";
 import {
   companyDataNames,
