+++ dataland-frontend/src/components/forms/CreateLksgDataset.vue
@@ -98,7 +98,7 @@ import { lksgDataModel } from "@/components/resources/frameworkDataSearch/lksg/L
 import { AxiosError } from "axios";
 import { CompanyAssociatedDataLksgData } from "@clients/backend";
 import { useRoute } from "vue-router";
-import { checkCustomInputs } from "@/utils/validationsUtils";
+import { checkCustomInputs } from "@/utils/ValidationsUtils";
 import NaceCodeFormField from "@/components/forms/parts/fields/NaceCodeFormField.vue";
 import InputTextFormField from "@/components/forms/parts/fields/InputTextFormField.vue";
 import FreeTextFormField from "@/components/forms/parts/fields/FreeTextFormField.vue";
@@ -112,8 +112,8 @@ import SubmitButton from "@/components/forms/parts/SubmitButton.vue";
 import SubmitSideBar from "@/components/forms/parts/SubmitSideBar.vue";
 import YesNoNaFormField from "@/components/forms/parts/fields/YesNoNaFormField.vue";
 import ProductionSiteFormField from "@/components/forms/parts/fields/ProductionSiteFormField.vue";
-import { objectDropNull, ObjectType } from "@/utils/updateObjectUtils";
-import { smoothScroll } from "@/utils/smoothScroll";
+import { objectDropNull, ObjectType } from "@/utils/UpdateObjectUtils";
+import { smoothScroll } from "@/utils/SmoothScroll";
 
 export default defineComponent({
   setup() {
