+++ dataland-frontend/src/components/general/SessionDialog.vue
@@ -32,7 +32,7 @@ import {
 } from "@/utils/SessionTimeoutUtils";
 import Keycloak from "keycloak-js";
 import { TIME_DISTANCE_SET_INTERVAL_SESSION_CHECK_IN_MS } from "@/utils/Constants";
-import { useSharedSessionStateStore } from "@/stores/stores";
+import { useSharedSessionStateStore } from "@/stores/Stores";
 import { loginAndRedirectToSearchPage } from "@/utils/KeycloakUtils";
 import { assertDefined } from "@/utils/TypeScriptUtils";
