+++ dataland-frontend/src/components/pages/UploadEuTaxonomyDataForFinancials.vue
@@ -13,7 +13,7 @@
 <script lang="ts">
 import TheHeader from "@/components/generics/TheHeader.vue";
 import AuthenticationWrapper from "@/components/wrapper/AuthenticationWrapper.vue";
-import CreateEuTaxonomyForFinancials from "@/components/forms/CreateEUTaxonomyForFinancials.vue";
+import CreateEuTaxonomyForFinancials from "@/components/forms/CreateEuTaxonomyForFinancials.vue";
 import CompanyInformation from "@/components/pages/CompanyInformation.vue";
 import TheFooter from "@/components/general/TheFooter.vue";
 import BackButton from "@/components/general/BackButton.vue";
