+++ dataland-frontend/src/components/pages/WelcomeDataland.vue
@@ -17,7 +17,7 @@ import { NavigationFailure, useRoute } from "vue-router";
 import Keycloak from "keycloak-js";
 import { assertDefined } from "@/utils/TypeScriptUtils";
 import SessionDialog from "@/components/general/SessionDialog.vue";
-import { useSharedSessionStateStore } from "@/stores/stores";
+import { useSharedSessionStateStore } from "@/stores/Stores";
 import { SessionDialogMode } from "@/utils/SessionTimeoutUtils";
 
 export default defineComponent({
