+++ dataland-frontend/src/utils/SessionTimeoutUtils.ts
@@ -1,7 +1,7 @@
 import { TIME_DISTANCE_SET_INTERVAL_SESSION_CHECK_IN_MS } from "@/utils/Constants";
 import { loginAndRedirectToSearchPage, logoutAndRedirectToUri } from "@/utils/KeycloakUtils";
 import Keycloak from "keycloak-js";
-import { useSharedSessionStateStore } from "@/stores/stores";
+import { useSharedSessionStateStore } from "@/stores/Stores";
 
 export enum SessionDialogMode {
   SessionWarning,
