+++ dataland-frontend/src/utils/SmoothScroll.ts
